+++ Rubrik/Public/Get-RubrikVCD.ps1
@@ -31,10 +31,17 @@ function Get-RubrikVCD
       .EXAMPLE
       Get-RubrikVCD -Status 'Connected'
       This returns the vCD settings on the currently connected Rubrik cluster with the status of 'Connected'
+
+      .EXAMPLE
+      Get-RubrikVCD -DetailedObject
+      This returns the full set of settings of the vCD clusters on the currently connected Rubrik cluster
   #>
 
   [CmdletBinding()]
   Param(
+    #ID of the VCD Cluster to retrieve
+    [ValidateNotNullOrEmpty()]
+    [String]$Id,
     # vCD Cluster Name
     [ValidateNotNullOrEmpty()]
     [String]$Name,
@@ -44,6 +51,9 @@ function Get-RubrikVCD
     # vCD Cluster Status
     [ValidateSet('Disconnected', 'Refreshing','Connected','BadlyConfigured','Deleting','Remote')]
     [String]$Status,
+    # DetailedObject will retrieved the detailed VCD object, the default behavior of the API is to only retrieve a subset of the full VCD object unless we query directly by ID. Using this parameter does affect performance as more data will be retrieved and more API-queries will be performed.
+    [Parameter(ParameterSetName='Query')]
+    [Switch]$DetailedObject,
     # Rubrik server IP or FQDN
     [String]$Server = $global:RubrikConnection.server,
     # API version
@@ -80,7 +90,18 @@ function Get-RubrikVCD
     $result = Test-ReturnFormat -api $api -result $result -location $resources.Result
     $result = Test-FilterObject -filter ($resources.Filter) -result $result
 
-    return $result
+    # If the Get-RubrikVCD function has been called with the -DetailedObject parameter a separate API query will be performed if the initial query was not based on ID
+    if (($DetailedObject) -and (-not $PSBoundParameters.containskey('id'))) {
+      for ($i = 0; $i -lt @($result).Count; $i++) {
+        $Percentage = [int]($i/@($result).count*100)
+        Write-Progress -Activity "DetailedObject queries in Progress, $($i+1) out of $(@($result).count)" -Status "$Percentage% Complete:" -PercentComplete $Percentage
+        Get-RubrikVCD -id $result[$i].id
+      }
+    } else {
+      return $result
+    }
+
+    #return $result
 
   } # End of process
 } # End of function